attack/gain support
Lets you specify an attack/release duration which gives the note a *twang* effect, imitating a much more realistic instrument.

Checks for a `this.attack` property, if found, increases the gain linearly using `linearRampToValueAtTime`. Then uses `this.release` (if found, or re-uses `this.attack`) to set gain back to 0.

If `this.attack` is not found, falls back to old behavior.

Usage:

    var note = notePlayer.buildFromName('C4');
    note.attack = 0.1;
    note.release = 0.5;
    note.play()